engine: extract esm/tours/selection.js
